added 5 quiz questions and its related options and correct answer

+++ src/services/QuestionAddAndRemoveService.java
@@ -1,22 +1,44 @@
 package services;
+import exceptions.NotAuthorizedToAddQuestionException;
+import exceptions.NotAuthorizedToRemoveQuestionException;
 import models.Options;
 import models.Questions;
+import models.Roles;
+import models.Users;
 import respositories.OptionsRepository;
 import respositories.QuestionsRepository;
+import respositories.UserRepository;
 
 import java.util.List;
 
 public class QuestionAddAndRemoveService {
     private QuestionsRepository questionsRepository;
     private OptionsRepository optionsRepository;
+    private UserRepository userRepository;
 
     public QuestionAddAndRemoveService(QuestionsRepository questionsRepository,
-                                       OptionsRepository optionsRepository) {
+                                       OptionsRepository optionsRepository,
+                                       UserRepository userRepository) {
         this.questionsRepository = questionsRepository;
         this.optionsRepository = optionsRepository;
+        this.userRepository = userRepository;
     }
 
-    public void addQuestionAndOptions(Questions questions) {
+    public void addQuestionAndOptions(Questions questions, String userName) throws NotAuthorizedToAddQuestionException {
+        Users users = userRepository.getUserFromDBUsingUserName(userName);
+
+        boolean canAdd = false;
+
+        for(Roles roles : users.getRoles()) {
+            if(roles.getRoleName().equals("ADMIN")) {
+                canAdd = true;
+            }
+        }
+
+        if(!canAdd) {
+            //the current user cannot add question
+            throw new NotAuthorizedToAddQuestionException(userName+" is not authorized to add question.");
+        }
         if(questionsRepository.getQuestionFromDBUsingQuestion(questions.getQuestion()) != null) {
             //Question is already present in DB
         }
@@ -34,7 +56,22 @@ public class QuestionAddAndRemoveService {
         }
     }
 
-    public void removeQuestion(String question) {
+    public void removeQuestion(String question, String userName) throws NotAuthorizedToRemoveQuestionException {
+        Users users = userRepository.getUserFromDBUsingUserName(userName);
+
+        boolean canRemove = false;
+
+        for(Roles roles : users.getRoles()) {
+            if(roles.getRoleName().equals("ADMIN")) {
+                canRemove = true;
+            }
+        }
+
+        if(!canRemove) {
+            //the current user cannot add question
+            throw new NotAuthorizedToRemoveQuestionException(userName+" is not authorized to remove question.");
+        }
+
         questionsRepository.removeQuestionFromDB(question);
     }
 }
